feat: Add party experience reward to PyBoyEnv
This commit adds the `party_exp_reward` attribute to the `PyBoyEnv` class in order to track the reward earned from the experience points (exp) gained by the player's party of Pokémon. The commit also includes modifications to the `reset()` method to reset the `party_exp_reward` and `party_exp` attributes to their initial values.